+++ src/imio/smartweb/core/browser/static/edit.js
@@ -1,13 +1,21 @@
 jQuery(document).ready(function ($) {
 
   // Toggle on add / edit forms fieldsets
-  // 1. hide default fieldset legend
+
+  // 1. hide default fieldset legend (ckass : expanded)
   $("form.tabbed-form-with-toggle fieldset:first legend").hide();
-  // 2. hide all fieldsets content except first
+  $("form.tabbed-form-with-toggle fieldset:first legend").addClass("expanded");
+
+  // 2. hide all fieldsets content except first (class : collapsed)
   $("form.tabbed-form-with-toggle fieldset:not(:first) legend").siblings().hide();
-  // 3. add toggle on all fieldsets legends
+  $("form.tabbed-form-with-toggle fieldset:not(:first) legend").addClass("collapsed");
+
+  // 3. add toggle on all fieldsets legends & toggle expanded / collapsed classes
   $("form.tabbed-form-with-toggle fieldset:not(:first) legend").click(function(){
-     $(this).siblings().slideToggle("fast");
+     var legend = $(this);
+     $(this).siblings().slideToggle("fast", function() {
+        legend.toggleClass("collapsed expanded");
+     });
   });
 
 });